+++ DatadogSessionReplay/Sources/Feature/SessionReplayFeature.swift
@@ -13,6 +13,7 @@ internal class SessionReplayFeature: SessionReplayConfiguration, DatadogRemoteFe
     let messageReceiver: FeatureMessageReceiver
     let performanceOverride: PerformancePresetOverride?
     let privacyLevel: SessionReplayPrivacyLevel
+    let imagePrivacyLevel: ImagePrivacyLevel
 
     // MARK: - Main Components
 
@@ -49,6 +50,7 @@ internal class SessionReplayFeature: SessionReplayConfiguration, DatadogRemoteFe
         let contextReceiver = RUMContextReceiver()
 
         self.privacyLevel = configuration.defaultPrivacyLevel
+        self.imagePrivacyLevel = configuration.defaultImagePrivacyLevel
         self.messageReceiver = CombinedFeatureMessageReceiver([
             contextReceiver,
             WebViewRecordReceiver(
@@ -59,6 +61,7 @@ internal class SessionReplayFeature: SessionReplayConfiguration, DatadogRemoteFe
         self.recordingCoordinator = RecordingCoordinator(
             scheduler: scheduler,
             privacy: configuration.defaultPrivacyLevel,
+            imagePrivacy: configuration.defaultImagePrivacyLevel,
             rumContextObserver: contextReceiver,
             srContextPublisher: SRContextPublisher(core: core),
             recorder: recorder,