+++ DatadogSessionReplay/Sources/Recorder/RecordingCoordinator.swift
@@ -15,6 +15,7 @@ internal class RecordingCoordinator {
     let scheduler: Scheduler
     let sampler: Sampler
     let privacy: PrivacyLevel
+    let imagePrivacy: ImagePrivacyLevel
     let srContextPublisher: SRContextPublisher
 
     private var currentRUMContext: RUMContext? = nil
@@ -28,6 +29,7 @@ internal class RecordingCoordinator {
     init(
         scheduler: Scheduler,
         privacy: PrivacyLevel,
+        imagePrivacy: ImagePrivacyLevel,
         rumContextObserver: RUMContextObserver,
         srContextPublisher: SRContextPublisher,
         recorder: Recording,
@@ -39,6 +41,7 @@ internal class RecordingCoordinator {
         self.scheduler = scheduler
         self.sampler = sampler
         self.privacy = privacy
+        self.imagePrivacy = imagePrivacy
         self.srContextPublisher = srContextPublisher
         self.telemetry = telemetry
         self.methodCallTelemetrySamplingRate = methodCallTelemetrySamplingRate
@@ -77,6 +80,7 @@ internal class RecordingCoordinator {
 
         let recorderContext = Recorder.Context(
             privacy: privacy,
+            imagePrivacy: imagePrivacy,
             applicationID: rumContext.applicationID,
             sessionID: rumContext.sessionID,
             viewID: viewID,